[ean13] 5 digit add-on implemented by direct calculation in GSUB
I initially thought calculating the check digit that encodes the pattern of
set-a set-b symbols was too complex to do in GSUB. But, I had a fancy idea
that reduces complexity a lot: of each calculation step, only use the least
significant digit (= result mod 10) of the result. That actually works
in this context (I checked it) and that makes it much better to implement
with GSUB, to know that each of the operations only produces one digit.